move swizzle to kernel
